requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
email_validator
werkzeug
pymysql
